fix: deferred codebase-analyst findings (issue #175)
- pain GET /: add from/to date filtering (parity with diary route)
- schemas: add .max(50) to chatRange and lastRange to reject oversized values
- env: assert COOKIE_SECURE=true in production at startup
- use-dashboard: pass from/to params to diary/pain queries; add unfiltered
  queries for global all-time streak unaffected by date range selection
- memorable-days: extract MemorableDayModal + EmojiPicker into its own
  component (memorable-day-modal.tsx); move shared CSS constants to
  memorable-days-constants.ts to fix props drilling and reduce file size

Closes #175
